Refactor: Separate main file into modular files and add documentation comments
