+++ backend/src/user/schemas.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class UserAuthSchema(BaseModel):
+    username: str
+    password: str
