+++ backend/tests/integration/test_news_endpoint.py
@@ -4,10 +4,10 @@ from sqlalchemy import create_engine, StaticPool
 from sqlalchemy.orm import sessionmaker
 import json
 from jose import jwt
-from main import app
-from main import Base, NewsArticle, User, session_opener, user_news_association_table
-from main import NewsSumaryRequestSchema, PromptRequest
-from main import pwd_context
+from backend.main import app
+from backend.main import Base, NewsArticle, User, session_opener, user_news_association_table
+from backend.main import NewsSumaryRequestSchema, PromptRequest
+from backend.main import pwd_context
 from unittest.mock import Mock
