+++ backend/tests/integration/test_price_endpoint.py
@@ -1,7 +1,7 @@
 import pytest
 from fastapi.testclient import TestClient
 from unittest.mock import patch
-from main import app
+from backend.main import app
 
 client = TestClient(app)
