+++ docker-compose.yml
@@ -17,7 +17,7 @@ services:
     volumes:
       - ./backend:/app
     environment:
-      - MODULE_NAME=app.main
+      - MODULE_NAME=app.src.main
       - VARIABLE_NAME=app
-    command: uvicorn main:app --host 0.0.0.0 --port 8000 --reload
+    command: uvicorn src.main:app --host 0.0.0.0 --port 8000 --reload
